Get snippets test to pass
@tooky I cheated a bit here. The output now contains both a Ruby snippet
and a wire snippet which was failing the test. It looks like this:

```
You can implement step definitions for undefined steps with these
snippets:

Given(/^we're all wired$/) do
  pending # Write code here that turns the phrase above into
concrete actions
end
foo()
  bar;
baz
```

I'm not sure why it used to not print Ruby snippets too, but it seems
more right that it does now, so I'm happy enough with this change in
behaviour. You too?

I don't like the output - I think there should be a newline between the
Ruby and wire snippets, and ideally we could change the protocol for
generators to give us some kind of heading name to explain what kind
of snippet it is, but anyway that's a different issue.