Tag final failing scenario as TODO
Nearly there!

@tooky this one is interesting. It's failing because of [this
code](https://github.com/cucumber/cucumber-ruby/blob/remove-wire-protocol-to-plugin/lib/cucumber/formatter/usage.rb#L24)
which expects to be able to query for step definitions on the fly using
the runtime 'god object' that was part of the original formatter API.

You can't get to wire protocol step defs like that anymore.

I'm tempted to just retire the usage formatter altogether, but maybe
there's a way we can store enough information about the match on the
core test_step (as the steps are activated) so that the usage formatter
has something to print at this point.

I've run out of steam / time here so haven't had a chance to investigate.